Rework of the downloader so it can deal with both types of pod caster pages. It now always uses the `https://www.podbean.com/site/EpisodeDownload/{slug}` pages and uses that to find the download links.
Had to rework the scraping method so it finds the correct page. It now
tries to find urls with the
`https://www.podbean.com/site/EpisodeDownload/{slug}` and then uses that
to find the rest of the podcast download links.

When looping back though pages. There is no 404 from the pages when you
find a page that isn't know. So we have to compare the HTML to the
previous page that we scraped. If they are the same we have probably
found the end.

Added debug and info logging to aid with development.

General tidy up of code and better organisation of functions so that
functions are logically grouped around the functions where they are
used.